Revert "rename read method names."
This reverts commit 219553e3999165a3787b34c3bf7f48c58223e359.
